pyproject.toml
[project]
name = "ComfyUI-Watermark-Detection-YOLO"
version = "1.0.0"
description = "Custom watermark detection using trained YOLO11 model for ComfyUI"
authors = [
    {name = "soliton80", email = "your-email@example.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
keywords = ["comfyui", "watermark", "detection", "yolo11", "custom-node"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Multimedia :: Graphics :: Graphics Conversion",
]
dependencies = [
    "ultralytics>=8.3.0",
    "opencv-python>=4.8.0",
    "pillow>=9.0.0",
    "numpy>=1.21.0",
]

[project.urls]
Homepage = "https://github.com/soliton80/ComfyUI-Watermark-Detection-YOLO"
Repository = "https://github.com/soliton80/ComfyUI-Watermark-Detection-YOLO"
Issues = "https://github.com/soliton80/ComfyUI-Watermark-Detection-YOLO/issues"
Documentation = "https://github.com/soliton80/ComfyUI-Watermark-Detection-YOLO/blob/main/README.md"

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"
